requirements.txt
langchain>=0.3.0
langchain-classic>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.7.4
streamlit>=1.40.0
python-dotenv>=1.0.1
sentence-transformers>=2.7.0
unstructured>=0.18.0
pdfplumber>=0.11.0
rank-bm25>=0.2.2
pydub>=0.25.1
requests>=2.31.0
streamlit-audio-recorder>=0.0.8

